Implementar y recolocar las tablas de la base de datos
